Add `importorskip` guard for `test_nvjpeg`
Local execution of `test_nvjpeg` will fail if `pynvjpeg` is not
available. This will result in test failures on any local machine that
does not have a GPU, or is otherwise not able to build and install
`pynvjpeg`. This PR adds an `importorskip` guard to skip the test when
`pynvjpeg` is not available.