Add `importorskip` guard for `test_nvjpeg` (#69)
* Add `importorskip` guard for `test_nvjpeg`

Local execution of `test_nvjpeg` will fail if `pynvjpeg` is not
available. This will result in test failures on any local machine that
does not have a GPU, or is otherwise not able to build and install
`pynvjpeg`. This PR adds an `importorskip` guard to skip the test when
`pynvjpeg` is not available.

* Add `is_mlo_like` / `is_cc_like` helpers (#70)

Adds helpers to `ViewPosition` and `MammogramView` that check if a view
is MLO-like or CC-like. Also adds an `is_standard_view` helper to
`ViewPosition` and adds tests for previously uncovered `MammogramView`
methods.